requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
attrs=19.1.0=pypi_0
autopep8=1.4.4=py_0
ca-certificates=2019.5.15=1
certifi=2019.6.16=py37_1
django=2.2.4=pypi_0
django-cors-headers=3.1.0=pypi_0
djangorestframework=3.10.2=pypi_0
jsonschema=3.0.2=pypi_0
openssl=1.1.1c=he774522_1
pep8=1.7.1=pypi_0
pip=19.1.1=py37_0
pycodestyle=2.5.0=py37_0
pyrsistent=0.15.4=pypi_0
python=3.7.4=h5263a28_0
pytz=2019.2=pypi_0
setuptools=41.0.1=py37_0
six=1.12.0=pypi_0
sqlite=3.29.0=he774522_0
sqlparse=0.3.0=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.15.26706=h3a45250_4
wheel=0.33.4=py37_0
wincertstore=0.2=py37_0
